CloudSQLReplicator/requirements.txt
absl-py>=0.6.1'
future>=0.17.1'
futures>=3.2.0'
google-api-core>=1.6.0'
google-api-python-client>=1.7.6'
google-auth>=1.6.1'
google-auth-httplib2>=0.0.3'
google-cloud-core>=0.28.1'
google-cloud-storage>=1.13.0'
google-resumable-media>=0.3.1'
googleapis-common-protos>=1.5.5'
httplib2>=0.12.0'
mysql-connector>=2.1.6'
oauth2client>=4.1.3'
six>=1.11.0'
